mod rest.rs

+++ src/rest.rs
@@ -58,7 +58,7 @@ const CONF_FINAL: usize = 10; // reorgs deeper than this are considered unlikely
 struct BlockValue {
     id: String,
     height: u32,
-    version: i32,
+    version: u32,
     timestamp: u32,
     tx_count: u32,
     size: u32,
@@ -86,7 +86,7 @@ impl BlockValue {
         BlockValue {
             id: header.block_hash().to_hex(),
             height: blockhm.header_entry.height() as u32,
-            version: header.version,
+            version: header.version as u32,
             timestamp: header.time,
             tx_count: blockhm.meta.tx_count,
             size: blockhm.meta.size,
@@ -115,7 +115,7 @@ impl BlockValue {
 #[derive(Serialize, Deserialize)]
 struct TransactionValue {
     txid: Txid,
-    version: i32,
+    version: u32,
     locktime: u32,
     vin: Vec<TxInValue>,
     vout: Vec<TxOutValue>,
@@ -152,7 +152,7 @@ impl TransactionValue {
 
         TransactionValue {
             txid: tx.txid(),
-            version: tx.version,
+            version: tx.version as u32,
             locktime: tx.lock_time,
             vin: vins,
             vout: vouts,
